update pinout of QFN and make it compilable

+++ common/ud3core/SignalGeneratorMIDI.c
@@ -33,8 +33,8 @@
 
 void synthcode_MIDI(uint32_t r){
     uint32_t rnd = rand();
-    uint8_t flag;
-	for (uint8_t ch = 0; ch < N_CHANNEL; ch++) {
+    uint32_t flag;
+	for (uint32_t ch = 0; ch < N_CHANNEL; ch++) {
         flag=0;
         if(channel[ch].volume > 0 && channel[ch].freq){
             if (Midi_voice[ch].noiseCurrent && (r / channel[ch].halfcount) % 2 > 0) {
